Change description of Room queries to Dao classname
